Move some of the custom convert struct to yaml.hpp
